Provide extra information
The existing device information class did not include
a lot of information that maybe useful: power-mode,
wifi-mac, ethernet-mac, etc.  So, instead of returning
a DeviceInfo instance by default, return a dictionary
containing the full results of the device-info query.

A class method has been added to the DeviceInfo class
to allow the creation of DeviceInfo instances from the
dictionaries now returned by the device_info property
of the Roku class.

As this change provides access to the 'power-mode'
field of the device-info result, add a `power_mode`
method to the Roku class to check for device power.